Add files via upload
This is my solution to problem #32 on practicepython.org (Create a hangman game)  The problem suggests drawing from a list of all words but I decided to only use names of common animals to make winning a little easier. :)  I added some validation of the user's guesses and tried to use functions wherever possible.
I'm sure this could have been done with less code, but I'm still relatively new to Python...I'd like to stop back and update this as I learn to be more efficient!

To play, you'll need to grab the wordlist_for_hangman_game.txt file from this same repository and place it in the same folder as this program.