Add files via upload
Data structure problem from the "#100DaysofCode in Python" program from the "Talk Python to Me" Podcast
To use the main program, you will also need to load the states_data.py file from this same repo in the same folder as the program

Perform the following tasks on the list and dict. The less you look at them, the better this exercise will be. Remember: Dicts are unsorted.

Print out the 10th item in each.
Print out the 45th key in the dictionary.
Print out the 27th value in the dictionary.